Move some helper from `HTMLChecker` to base class
