+++ src/checkers/htmlchecker.py
@@ -21,7 +21,6 @@
 import logging
 import re
 import urllib.parse
-from string import Template
 from distutils.version import LooseVersion
 import io
 import codecs
@@ -30,13 +29,7 @@ import typing as t
 import aiohttp
 from yarl import URL
 
-from ..lib import (
-    utils,
-    NETWORK_ERRORS,
-    WRONG_CONTENT_TYPES_FILE,
-    WRONG_CONTENT_TYPES_ARCHIVE,
-    FILE_URL_SCHEMES,
-)
+from ..lib import NETWORK_ERRORS
 from ..lib.externaldata import ExternalBase, ExternalData
 from ..lib.errors import CheckerMetadataError, CheckerQueryError, CheckerFetchError
 from ..lib.checkers import Checker
@@ -183,56 +176,3 @@ class HTMLChecker(Checker):
         abs_url = urllib.parse.urljoin(base=url, url=latest_url)
 
         await self._update_version(external_data, latest_version, abs_url)
-
-    @staticmethod
-    def _substitute_placeholders(template_string: str, version: str) -> str:
-        version_list = LooseVersion(version).version
-        tmpl = Template(template_string)
-        tmpl_vars: t.Dict[str, t.Union[str, int]]
-        tmpl_vars = {"version": version}
-        for i, version_part in enumerate(version_list):
-            tmpl_vars[f"version{i}"] = version_part
-            if i == 0:
-                tmpl_vars["major"] = version_part
-            elif i == 1:
-                tmpl_vars["minor"] = version_part
-            elif i == 2:
-                tmpl_vars["patch"] = version_part
-        try:
-            return tmpl.substitute(**tmpl_vars)
-        except (KeyError, ValueError) as err:
-            raise CheckerMetadataError("Error substituting template") from err
-
-    async def _update_version(
-        self,
-        external_data: ExternalData,
-        latest_version: str,
-        latest_url: str,
-        follow_redirects: bool = False,
-    ):
-        assert latest_version is not None
-        assert latest_url is not None
-
-        if external_data.type == ExternalData.Type.ARCHIVE:
-            wrong_content_types = WRONG_CONTENT_TYPES_ARCHIVE
-        else:
-            wrong_content_types = WRONG_CONTENT_TYPES_FILE
-
-        latest_url_scheme = URL(latest_url).scheme
-        if latest_url_scheme not in FILE_URL_SCHEMES:
-            raise CheckerMetadataError(f"Invalid URL scheme {latest_url_scheme}")
-
-        try:
-            new_version = await utils.get_extra_data_info_from_url(
-                url=latest_url,
-                follow_redirects=follow_redirects,
-                session=self.session,
-                content_type_deny=wrong_content_types,
-            )
-        except NETWORK_ERRORS as err:
-            raise CheckerFetchError from err
-
-        new_version = new_version._replace(
-            version=latest_version  # pylint: disable=no-member
-        )
-        external_data.set_new_version(new_version)
