+++ src/lib/checkers.py
@@ -1,11 +1,19 @@
 import logging
+from distutils.version import LooseVersion
+from string import Template
 import typing as t
 
 import aiohttp
 from yarl import URL
 import jsonschema
 
-from . import NETWORK_ERRORS
+from . import (
+    utils,
+    NETWORK_ERRORS,
+    WRONG_CONTENT_TYPES_FILE,
+    WRONG_CONTENT_TYPES_ARCHIVE,
+    FILE_URL_SCHEMES,
+)
 from .externaldata import (
     ExternalBase,
     ExternalData,
@@ -69,6 +77,25 @@ class Checker:
 
     # Various helplers for checkers; assumed to be safely usable only from subclasses
 
+    @staticmethod
+    def _substitute_placeholders(template_string: str, version: str) -> str:
+        version_list = LooseVersion(version).version
+        tmpl = Template(template_string)
+        tmpl_vars: t.Dict[str, t.Union[str, int]]
+        tmpl_vars = {"version": version}
+        for i, version_part in enumerate(version_list):
+            tmpl_vars[f"version{i}"] = version_part
+            if i == 0:
+                tmpl_vars["major"] = version_part
+            elif i == 1:
+                tmpl_vars["minor"] = version_part
+            elif i == 2:
+                tmpl_vars["patch"] = version_part
+        try:
+            return tmpl.substitute(**tmpl_vars)
+        except (KeyError, ValueError) as err:
+            raise CheckerMetadataError("Error substituting template") from err
+
     async def _complete_digests(
         self, url: t.Union[str, URL], digests: MultiDigest
     ) -> MultiDigest:
@@ -113,3 +140,37 @@ class Checker:
             new_version = new_version._replace(checksum=checksum)
 
         source.set_new_version(new_version)
+
+    async def _update_version(
+        self,
+        external_data: ExternalData,
+        latest_version: str,
+        latest_url: str,
+        follow_redirects: bool = False,
+    ):
+        assert latest_version is not None
+        assert latest_url is not None
+
+        if external_data.type == ExternalData.Type.ARCHIVE:
+            wrong_content_types = WRONG_CONTENT_TYPES_ARCHIVE
+        else:
+            wrong_content_types = WRONG_CONTENT_TYPES_FILE
+
+        latest_url_scheme = URL(latest_url).scheme
+        if latest_url_scheme not in FILE_URL_SCHEMES:
+            raise CheckerMetadataError(f"Invalid URL scheme {latest_url_scheme}")
+
+        try:
+            new_version = await utils.get_extra_data_info_from_url(
+                url=latest_url,
+                follow_redirects=follow_redirects,
+                session=self.session,
+                content_type_deny=wrong_content_types,
+            )
+        except NETWORK_ERRORS as err:
+            raise CheckerFetchError from err
+
+        new_version = new_version._replace(
+            version=latest_version  # pylint: disable=no-member
+        )
+        external_data.set_new_version(new_version)
